test: remove OAuth callback test file (part of removed feature)
This test file was added in commit 7ab4029 as part of the OAuth reauthentication
feature with GET browser redirect support. Since we're removing that feature and
reverting to POST-only (as it was on main), this test file should also be removed.

The POST OAuth callback functionality is already working in production and is
tested through integration tests. This complex unit test with extensive mocking
was specific to the GET redirect feature we removed.